Делает мелкие правки
