+++ src/exception/SourceFileException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Taskforce\Exception;
+namespace Taskforce\exception;
 
 class SourceFileException extends BaseException
 {
