+++ src/exception/StatusException.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Taskforce\Exception;
+namespace Taskforce\exception;
 
 class StatusException extends BaseException
 {
